requirements.txt
flask
pytest
google
beautifulsoup4
requests
newspaper3k
data_object